Support half-size rim sizes (e.g. 22.5) for truck tires
The rim column was an INT, so a 22.5" truck tire rim size silently
got truncated to 22. Changed it to DECIMAL(4,1) end-to-end (schema,
create/update/list APIs) and gave the admin form's rim field a 0.1
step. Whole sizes still display without a trailing ".0" since the
value is a plain number on the frontend.

Includes a one-time migrate_rim_decimal.php script for the
already-deployed database — run once in the browser, then delete,
same pattern as the other migrations.

+++ php-mysql-version/install.sql
@@ -23,7 +23,7 @@ CREATE TABLE IF NOT EXISTS products (
   model       VARCHAR(120) NOT NULL,
   width       INT NULL,
   profile     INT NULL,
-  rim         INT NOT NULL,
+  rim         DECIMAL(4,1) NOT NULL,
   hole_count  INT NULL,
   pcd         VARCHAR(10) NULL,
   season      ENUM('summer','winter','all-season') NULL,
